requirements.txt
openpyxl
snowflake-connector-python
requests
